+++ src/components/kitchen/index.ts
@@ -17,3 +17,5 @@ export { PackingPanel } from './PackingPanel';
 export { FlipNumber, FlowChevrons } from './FlipNumber';
 export { GoFoodStickerCard } from './GoFoodStickerCard';
 export { GoFoodPackingCard } from './GoFoodPackingCard';
+export { K3MartStockCard } from './K3MartStockCard';
+export { K3MartPackingCard } from './K3MartPackingCard';